Feature/cesar

+++ src/org/generation/hackathon/models/Contacto.java
@@ -6,13 +6,13 @@ public class Contacto {
     private int id;
     private String nombre;
     private String apellido;
-    private int telefono;
+    private long telefono;
     private static int idContacto = 1;
 
     public Contacto() {
     }
 
-    public Contacto(String nombre, String apellido, int telefono) {
+    public Contacto(String nombre, String apellido, long telefono) {
         this.id = idContacto++;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -43,7 +43,7 @@ public class Contacto {
         this.apellido = apellido;
     }
 
-    public int getTelefono() {
+    public long getTelefono() {
         return telefono;
     }
 
@@ -63,8 +63,4 @@ public class Contacto {
     public String toString() {
         return "Contacto [id=" + id + ", nombre=" + nombre + ", apellido=" + apellido + ", telefono=" + telefono + "]";
     }
-    
-    
-
-
 }
